requirements.txt
asgiref==3.4.1
click==8.0.1
colorama==0.4.4
coverage==5.5
fastapi==0.68.1
h11==0.12.0
mysql-connector-python==8.0.26
protobuf==3.17.3
pydantic==1.8.2
python-dotenv==0.19.0
six==1.16.0
starlette==0.14.2
typing-extensions==3.10.0.1
uvicorn==0.15.0
